updated tools to work with internal Jnomics file format

+++ tools/src/main/java/edu/cshl/schatz/jnomics/tools/Bowtie2.java
@@ -1,33 +1,31 @@
 package edu.cshl.schatz.jnomics.tools;
 
-import java.io.*;
-import java.util.Arrays;
-import java.util.Iterator;
-import java.util.List;
-
-import edu.cshl.schatz.jnomics.io.*;
+import edu.cshl.schatz.jnomics.cli.OptionBuilder;
+import edu.cshl.schatz.jnomics.io.ThreadedStreamConnector;
+import edu.cshl.schatz.jnomics.mapreduce.JnomicsTool;
 import edu.cshl.schatz.jnomics.ob.ReadCollectionWritable;
-import edu.cshl.schatz.jnomics.ob.ReadWritable;
-import edu.cshl.schatz.jnomics.util.Functional.Operation;
-import edu.cshl.schatz.jnomics.util.TextUtil;
+import edu.cshl.schatz.jnomics.ob.SAMRecordWritable;
+import net.sf.samtools.SAMFileReader;
+import net.sf.samtools.SAMRecord;
+import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.Option;
 import org.apache.commons.cli.Options;
 import org.apache.commons.cli.ParseException;
-import org.apache.commons.cli.CommandLine;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.filecache.DistributedCache;
 import org.apache.hadoop.io.NullWritable;
-import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Job;
 import org.apache.hadoop.mapreduce.Mapper;
+import org.apache.hadoop.mapreduce.Reducer;
 import org.apache.hadoop.mapreduce.lib.input.SequenceFileInputFormat;
-import org.apache.hadoop.mapreduce.lib.output.TextOutputFormat;
+import org.apache.hadoop.mapreduce.lib.output.SequenceFileOutputFormat;
 
-import edu.cshl.schatz.jnomics.cli.OptionBuilder;
-import edu.cshl.schatz.jnomics.mapreduce.JnomicsTool;
-import edu.cshl.schatz.jnomics.ob.writable.QueryTemplate;
+import java.io.BufferedOutputStream;
+import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
 
 public class Bowtie2 extends JnomicsTool{
 
@@ -48,26 +46,28 @@ public class Bowtie2 extends JnomicsTool{
 	public int run(String[] args) throws Exception {
 		
         Configuration configuration = getConf();
-		Job job = new Job(configuration);
-				
-		DistributedCache.createSymlink(configuration);
-				
-		job.setMapperClass(Bowtie2.Bowtie2Map.class);
+        Job job = new Job(configuration);
 
-		job.setInputFormatClass(SequenceFileInputFormat.class);
-		job.setOutputFormatClass(TextOutputFormat.class);
-		
-		job.setMapOutputKeyClass(Text.class);
+		DistributedCache.createSymlink(configuration);
+
+        job.setMapperClass(Bowtie2.Bowtie2Map.class);
+        job.setReducerClass(Reducer.class);
+
+        job.setInputFormatClass(SequenceFileInputFormat.class);
+		job.setOutputFormatClass(SequenceFileOutputFormat.class);
+
+		job.setMapOutputKeyClass(SAMRecordWritable.class);
 		job.setMapOutputValueClass(NullWritable.class);
-        job.setOutputKeyClass(Text.class);
+        job.setOutputKeyClass(SAMRecordWritable.class);
         job.setOutputValueClass(NullWritable.class);
-        
+
         job.setNumReduceTasks(0);
-		job.setJarByClass(Bowtie2.class);
-				
+
+        job.setJarByClass(Bowtie2.class);
+
 		return job.waitForCompletion(true) ? 0 : 1;
     }
-	
+    
 	@Override
 	public Options getOptions() {
        Options options = super.getOptions();
@@ -122,12 +122,13 @@ public class Bowtie2 extends JnomicsTool{
         return STATUS_OK;
 	}
 
-    public static class Bowtie2Map extends Mapper<ReadCollectionWritable,NullWritable,Text,NullWritable>{
+    public static class Bowtie2Map extends Mapper<ReadCollectionWritable,NullWritable,SAMRecordWritable,NullWritable> {
 
 		private File[] tmpFiles;
 		private Thread readerThread, bowtieProcessErrThread;
 		private Process bowtieProcess;
         private String cmd;
+        private Throwable readerError = null;
 
         @Override
 		protected void setup(final Context context) throws IOException,
@@ -149,18 +150,15 @@ public class Bowtie2 extends JnomicsTool{
                 f.createNewFile();
                 f.deleteOnExit();
             }
-           
-            
+
             cmd = String.format(
                     "%s %s --mm -x %s -1 %s -2 %s",
                     bowtie_binary, bowtie_opts, bowtie_idx,
         			tmpFiles[0],
         			tmpFiles[1]);
-            
 
 		}
 
-		
 		@Override
 		public void run(final Context context) throws IOException, InterruptedException {
 			setup(context);
@@ -183,19 +181,25 @@ public class Bowtie2 extends JnomicsTool{
             bowtieProcessErrThread = new Thread(new ThreadedStreamConnector(bowtieProcess.getErrorStream(), System.err));
             bowtieProcessErrThread.start();
 
-            /** Read lines from sampe stdout and print them to context **/
-            readerThread = new Thread(new ThreadedLineOperator(bowtieProcess.getInputStream(), new Operation(){
-                private final Text line = new Text();
+            /** Read lines from bowtie stdout and print them to context **/
+            readerThread = new Thread( new Runnable() {
+
+                private final SAMRecordWritable writableRecord = new SAMRecordWritable();
+                
                 @Override
-                public <T> void performOperation(T data){
-                    line.set((String)data);
-                    try{
-                        context.write(line,NullWritable.get());
-                    }catch(Exception e){
-                        assert(false);
+                public void run() {
+                    SAMFileReader reader = new SAMFileReader(bowtieProcess.getInputStream());
+                    for(SAMRecord record: reader){
+                        writableRecord.set(record);
+                        try {
+                            context.write(writableRecord,NullWritable.get());
+                        }catch(Exception e){
+                            readerError = e;
+                        }
                     }
+                    reader.close();
                 }
-            }));
+            });
 
             readerThread.start();
 
@@ -208,13 +212,17 @@ public class Bowtie2 extends JnomicsTool{
 				InterruptedException {
 			
             readerThread.join();
+            if(readerError != null)
+                throw new IOException(readerError);
+
             bowtieProcess.waitFor();
             bowtieProcessErrThread.join();
+
             for(File f: tmpFiles){
                 f.delete();
             }
 		}
-						
+
 	}
 			
 }
